Simplify and improve tests
