+++ django_workflow_engine/tests/test_workflows/test_linear_workflow.py
@@ -1,4 +1,5 @@
 import pytest
+from django_workflow_engine.models import TaskRecord
 from django_workflow_engine.tests.utils import set_up_flow
 from django_workflow_engine.tests.workflows import linear_workflow
 
@@ -11,7 +12,13 @@ def test_workflow_creation(settings):
     )
     executor.run_flow(user=test_user)
 
+    assert TaskRecord.objects.count() == 3
 
-def test_only_group_member_can_execute():
-    # TODO
-    pass
+    correct_task_order = [
+        "start",
+        "task_a",
+        "task_b",
+    ]
+
+    for i, task_record in enumerate(TaskRecord.objects.all()):
+        assert task_record.step_id == correct_task_order[i]
