+++ tox.ini
@@ -6,5 +6,6 @@ deps=
     django>=3.2.8
     pytest
     pytest-django
+    factory_boy
 commands=
     pytest
